refactor: Price 피드백, Auto Formatting  사용

+++ src/test/java/kitchenpos/common/PriceTest.java
@@ -4,6 +4,8 @@ import kitchenpos.menu.exception.WrongPriceException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -12,15 +14,15 @@ class PriceTest {
     @DisplayName("0보다 큰 수를 입력해야 한다.")
     @Test
     void bigThanZero() {
-        assertThatThrownBy(() -> new Price(-1))
+        assertThatThrownBy(() -> new Price(BigDecimal.valueOf(-1)))
                 .isInstanceOf(WrongPriceException.class);
     }
 
     @DisplayName("곱셈")
     @Test
     void multiply() {
-        Price price = new Price(100);
-        assertThat(price.multiply(25)).isEqualTo(new Price(2500));
+        Price price = new Price(BigDecimal.valueOf(100));
+        assertThat(price.multiply(25)).isEqualTo(new Price(BigDecimal.valueOf(2500)));
     }
 
 }
